feat: Separate logger
