+++ src/prelude.rs
@@ -10,9 +10,11 @@ pub use crate::{
 pub use crate::compat::adapt::FmtWriteAdapter;
 pub use crate::compat::ascii::{write_byte_hex, write_bytes_hex, write_bytes_hex_prefixed};
 pub use crate::compat::{HalErrorExt, I2cCompat, SerialCompat, SerialEio, UartLike};
-pub use crate::explorer::{CmdExecutor, CmdNode, ExecutorError, Explorer, NullLogger};
+pub use crate::explorer::{CmdExecutor, CmdNode, ExecutorError, Explorer};
 // In src/prelude.rs
 
 pub use crate::scanner::{
-    LogLevel, run_explorer, run_single_sequence_explorer, scan_i2c, scan_init_sequence,
+    run_explorer, run_single_sequence_explorer, scan_i2c, scan_init_sequence,
 };
+
+pub use crate::logger::{Logger, SerialLogger, LogLevel};
